Reviewing feats Stealth Synergy, Stalwart, Stage Combatant, Sneaking Precision, Sorcerous Strike*, Spell Bane, Spinning Throw, and Splintering Weapon*.
Sorcerous Strike:
This feat should have the user selecting one spell from their available list of Bloodline spells, but I'm not certain how to accomplish this. I'll have to research the CHOOSE tag a little more before attempting to implement this part of it.

Splintering Weapon:
This feat requires the user to select a fragile weapon, one they are proficient with. I have it set to only list fragile weapons, but regardless of what I'm doing, it's still listing weapons the character is not proficient with. Maybe I'm not using the QUALIFIED tag right? Need to find out.
Issue#: NEWSOURCE-130


git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk/content/notfordistribution@17145 ef53c05b-3124-4d59-ae7d-fe3273f42be7