Clean up abilities and class
Moved all of the bonuses affecting Arcane Reservoir and Arcanist
exploits from the class into the abilites controlling those
features.  Renamed the variables in the Arcane Reservoir to
better reflect what they are.

Changed the name of the Slayer Trapfinding feature to
"Trapfinding - Slayer" as it grants the Rogue ability Trapfinding
and they both end up on the character sheet.

Made the trapspotter actually grant the Disable Device class
skill.

Made the Warpriest Class Level for Sacred weapon independent of
actual class level so it can be bonused by favoured class.